requirements.txt
apache-airflow[celery]
xmltodict==0.13.0